remove plone.app.transmogrifier dependency
