+++ base.cfg
@@ -138,7 +138,4 @@ redturtle.importer.base =
 plone.restapi = 
 
 [sources]
-transmogrify.dexterity  = git https://github.com/collective/transmogrify.dexterity.git pushurl=git@github.com:collective/transmogrify.dexterity.git branch=eikichi18-python3-fix
 collective.transmogrifier = git https://github.com/collective/collective.transmogrifier.git pushurl=git@github.com:collective/collective.transmogrifier.git branch=eikichi18-python3-fix
-plone.app.transmogrifier = git https://github.com/collective/plone.app.transmogrifier.git pushurl=git@github.com:collective/plone.app.transmogrifier.git branch=python3
-collective.jsonmigrator = git https://github.com/collective/collective.jsonmigrator.git pushurl=git@github.com:collective/collective.jsonmigrator.git branch=eikichi18-python3-fix
