+++ src/redturtle/importer/base/transmogrifier/redturtleplone5.cfg
@@ -60,7 +60,7 @@ blueprint = redturtle.importer.base.sections.schemaupdater
 blueprint = redturtle.importer.base.sections.datafields
 
 [datesupdater]
-blueprint = plone.app.transmogrifier.datesupdater
+blueprint = redturtle.importer.base.sections.datesupdater
 path-key = _path
 creation-key = creation_date
 modification-key = modification_date
@@ -78,19 +78,19 @@ blueprint = redturtle.importer.base.sections.discussions
 path-key = _path
 
 [properties]
-blueprint = collective.jsonmigrator.properties
+blueprint = redturtle.importer.base.sections.properties
 
 [browser-default]
-blueprint = plone.app.transmogrifier.browserdefault
+blueprint = redturtle.importer.base.sections.browserdefault
 
 [owner]
-blueprint = collective.jsonmigrator.owner
+blueprint = redturtle.importer.base.sections.owner
 
 [local_roles]
 blueprint = redturtle.importer.base.sections.localroles
 
 [reindexobject]
-blueprint = plone.app.transmogrifier.reindexobject
+blueprint = redturtle.importer.base.sections.reindexobject
 
 [context_fixes]
 blueprint = redturtle.importer.base.sections.context_fixes
