+++ test_plone52.cfg
@@ -9,8 +9,9 @@ update-versions-file = test_plone52.cfg
 
 [versions]
 plone.testing = 7.0.1
-pyflakes = 1.5.0
 flake8 = 3.3.0
+pyflakes = 1.5.0
+pycodestyle = 2.0.0
 
 # Added by buildout at 2020-02-03 11:59:33.269044
 PyYAML = 5.3
